Fix: Split Writer test file
