+++ test/unit/domain/warp/Writer.test.ts
@@ -68,9 +68,9 @@ function createPatchMessage(lamport = 1) {
 }
 
 describe('Writer (WARP schema:2)', () => {
-    let persistence;
-    let versionVector;
-    let getCurrentState;
+  let persistence;
+  let versionVector;
+  let getCurrentState;
 
   beforeEach(() => {
     persistence = createMockPersistence();
@@ -689,151 +689,5 @@ describe('Writer (WARP schema:2)', () => {
       expect(err instanceof Error).toBe(true);
       expect(err instanceof WriterError).toBe(true);
     });
-  });
-});
-
-describe('PatchSession operations', () => {
-    let persistence;
-    let versionVector;
-    let getCurrentState;
-    let patchJournal;
-
-  beforeEach(() => {
-    persistence = createMockPersistence();
-    versionVector = VersionVector.empty();
-    getCurrentState = vi.fn(() => null);
-    persistence.readRef.mockResolvedValue(null);
-    patchJournal = createPatchJournal(persistence);
-  });
-
-  it('addNode creates node-add op', async () => {
-    const writer = new Writer({
-      persistence,
-      patchJournal,
-      graphName: 'events',
-      writerId: 'alice',
-      versionVector,
-      getCurrentState,
-    });
-
-    const patch = await writer.beginPatch();
-    patch.addNode('user:alice');
-
-    const built = (patch.build() as any);
-    expect(built.ops).toHaveLength(1);
-    expect(built.ops[0].type).toBe('NodeAdd');
-    expect(built.ops[0].node).toBe('user:alice');
-  });
-
-  it('removeNode creates node-remove op', async () => {
-    const state = ({ nodeAlive: ORSet.empty(), edgeAlive: ORSet.empty(), prop: new Map(), observedFrontier: VersionVector.empty() } as any);
-    state.nodeAlive.add('user:alice', Dot.create('alice', 1));
-
-    const writer = new Writer({
-      persistence,
-      patchJournal,
-      graphName: 'events',
-      writerId: 'alice',
-      versionVector,
-      getCurrentState: () => state,
-    });
-
-    const patch = await writer.beginPatch();
-    patch.removeNode('user:alice');
-
-    const built = (patch.build() as any);
-    expect(built.ops).toHaveLength(1);
-    expect(built.ops[0].type).toBe('NodeRemove');
-    expect(built.ops[0].node).toBe('user:alice');
-  });
-
-  it('addEdge creates edge-add op', async () => {
-    const writer = new Writer({
-      persistence,
-      patchJournal,
-      graphName: 'events',
-      writerId: 'alice',
-      versionVector,
-      getCurrentState,
-    });
-
-    const patch = await writer.beginPatch();
-    patch.addEdge('n1', 'n2', 'links');
-
-    const built = (patch.build() as any);
-    expect(built.ops).toHaveLength(1);
-    expect(built.ops[0].type).toBe('EdgeAdd');
-    expect(built.ops[0].from).toBe('n1');
-    expect(built.ops[0].to).toBe('n2');
-    expect(built.ops[0].label).toBe('links');
-  });
-
-  it('removeEdge creates edge-remove op', async () => {
-    const state = ({ nodeAlive: ORSet.empty(), edgeAlive: ORSet.empty(), prop: new Map(), observedFrontier: VersionVector.empty() } as any);
-    const ek = encodeEdgeKey('n1', 'n2', 'links');
-    state.edgeAlive.add(ek, Dot.create('alice', 1));
-
-    const writer = new Writer({
-      persistence,
-      patchJournal,
-      graphName: 'events',
-      writerId: 'alice',
-      versionVector,
-      getCurrentState: () => state,
-    });
-
-    const patch = await writer.beginPatch();
-    patch.removeEdge('n1', 'n2', 'links');
-
-    const built = (patch.build() as any);
-    expect(built.ops).toHaveLength(1);
-    expect(built.ops[0].type).toBe('EdgeRemove');
-  });
-
-  it('setProperty creates prop-set op', async () => {
-    const writer = new Writer({
-      persistence,
-      patchJournal,
-      graphName: 'events',
-      writerId: 'alice',
-      versionVector,
-      getCurrentState,
-    });
-
-    const patch = await writer.beginPatch();
-    patch.setProperty('user:alice', 'name', 'Alice');
-
-    const built = (patch.build() as any);
-    expect(built.ops).toHaveLength(1);
-    expect(built.ops[0].type).toBe('PropSet');
-    expect(built.ops[0].node).toBe('user:alice');
-    expect(built.ops[0].key).toBe('name');
-    expect(built.ops[0].value).toBe('Alice');
-  });
-
-  it('supports various property value types', async () => {
-    const writer = new Writer({
-      persistence,
-      patchJournal,
-      graphName: 'events',
-      writerId: 'alice',
-      versionVector,
-      getCurrentState,
-    });
-
-    const patch = await writer.beginPatch();
-    patch.setProperty('n', 'str', 'hello');
-    patch.setProperty('n', 'num', 42);
-    patch.setProperty('n', 'bool', true);
-    patch.setProperty('n', 'arr', [1, 2, 3]);
-    patch.setProperty('n', 'obj', { x: 1 });
-
-    const built = (patch.build() as any);
-    expect(built.ops).toHaveLength(5);
-    expect(built.ops[0].value).toBe('hello');
-    expect(built.ops[1].value).toBe(42);
-    expect(built.ops[2].value).toBe(true);
-    expect(built.ops[3].value).toEqual([1, 2, 3]);
-    expect(built.ops[4].value).toEqual({ x: 1 });
   });
 });
